added actionFlags; minor fix on projectile

+++ src/player/GOM_Projectile.cpp
@@ -3,7 +3,7 @@
 size_t GOM_Projectile::projectileCount_ = 0;
 
 // constructor
-GOM_Projectile::GOM_Projectile(int objID, int type, int degree, int posX, int posY, 
+GOM_Projectile::GOM_Projectile(int objID, int type, double degree, double posX, double posY, 
 	int playerID, int speed, int ttl, int damage)
 :GameObjectMoveable(objID, type, degree, posX, posY, playerID, speed), 
 	ttl_(ttl), damage_(damage)
